fix: address Clawpatch finding fnd_sig-feat-lib
